+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.x.y / 2014-10-xx
 ------------------
 - prerender for SEO
+- added `bitcoin:` anchor on checkout (so you can click and your client will open up)
 
 0.0.1 / 2014-10-09
 ------------------